backend/requirements.txt
aiofiles
Django
djangorestframework
django-rest-knox
django-cors-headers
dj-rest-auth
django-allauth
djangorestframework-simplejwt
djangorestframework-camel-case
Pillow
psycopg2-binary
python-environ
stream-zip
uvicorn[standard]
xlsxwriter